requirements.txt
pytest-mozwebqa
requests==2.4.3
